Revert "merge the stat and data records"
