--- tests/fs/make-fs-v1.html
@@ -1,72 +1,0 @@
-<!doctype html>
-<html>
-
-<head>
-  <meta charset="UTF-8">
-  <script type="text/javascript" src="../../timer.js"></script>
-  <script type="text/javascript" src="fs-v1.js"></script>
-</head>
-
-<body>
-<script type="text/javascript">
-  function load(file, responseType) {
-    return new Promise(function(resolve, reject) {
-      var xhr = new XMLHttpRequest();
-      xhr.open("GET", file, true);
-      xhr.responseType = responseType;
-      xhr.onload = function () {
-        resolve(xhr.response);
-      };
-      xhr.onerror = function() {
-        reject();
-      };
-      xhr.send(null);
-    });
-  }
-
-  new Promise(function(resolve, reject) {
-    fs.init(resolve);
-  }).then(function() {
-    return Promise.all([
-      new Promise(function(resolve, reject) {
-        fs.mkdir("/Persistent", resolve);
-      }),
-
-      new Promise(function(resolve, reject) {
-        fs.exists("/_main.ks", function(exists) {
-          if (exists) {
-            resolve();
-          } else {
-            load("../../certs/_main.ks", "blob").then(function(data) {
-              fs.create("/_main.ks", data, function() {
-                resolve();
-              });
-            });
-          }
-        });
-      }),
-
-      new Promise(function(resolve, reject) {
-        fs.exists("/_test.ks", function(exists) {
-          if (exists) {
-            resolve();
-          } else {
-            load("../../certs/_test.ks", "blob").then(function(data) {
-              fs.create("/_test.ks", data, function() {
-                resolve();
-              });
-            });
-          }
-        });
-      }),
-    ]);
-  }).then(function() {
-    return new Promise(function(resolve, reject) {
-      fs.storeSync(resolve);
-    });
-  }).then(function() {
-    document.body.appendChild(document.createTextNode("DONE"));
-  });
-</script>
-</body>
-</html>
