+++ tests/fstests.html
@@ -4,8 +4,8 @@
 <head>
   <meta charset="UTF-8">
   <meta name="viewport" content="initial-scale=1.0">
-  <script type="text/javascript" src="../../libs/fs.js" defer></script>
-  <script type="text/javascript" src="../../timer.js" defer></script>
+  <script type="text/javascript" src="../libs/fs.js" defer></script>
+  <script type="text/javascript" src="../timer.js" defer></script>
   <script type="text/javascript">
   console.log = function() {
     var s = Array.prototype.join.call(arguments, ",") +"\n";
